Consolidate redundant code: shared escape and date utilities
- Create src/utils/escapeUtils.ts with escapeHtml, escapeLatex, and
  escapeString functions, consolidating 8+ duplicate implementations
- Update 18 files to use shared dateConstants.ts for day/month names,
  eliminating 14+ duplicate ['Sun', 'Mon', ...] arrays
- Update orgRepeater.ts to import from dateConstants.ts
- Update orgExport.ts to import and re-export escapeString from escapeUtils
- Update orgLinkTypes.ts to import escape functions from escapeUtils
- Update publishing modules (orgPublish, defaultTheme, bookTheme/layout)
  to use shared escapeHtml
- Update tableProvider.ts to use shared escapeHtml
- Update audit report to reflect completed redundancy cleanup
